Add ConsensusVersion::V10 and move all related logic into consensus_heights.rs
